% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{billtype}
\alias{billtype}
\title{billtype map}
\format{
`billtype table`
a dataset to facilitate billtype translate-ability
}
\source{
https://med.noridianmedicare.com/web/jea/topics/claim-submission/bill-types
}
\usage{
billtype
}
\description{
a map for cms billtypes
}
\details{
as of Apr 30th 2023
}
\keyword{datasets}
