% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/billtype.R
\name{billtype_check}
\alias{billtype_check}
\title{Validate Billtype}
\usage{
billtype_check(btcode)
}
\arguments{
\item{btcode}{a billtype code or vector of codes}
}
\value{
boolean for validity
}
\description{
check if billtype code is valid
}
\examples{
billtype_check('0111')
}
