% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnosis.R
\name{diag_scrape}
\alias{diag_scrape}
\title{Full Webscrape of ICD-10 data}
\usage{
diag_scrape(zzz = 5)
}
\arguments{
\item{zzz}{how long inbetween website pulls... please don't go below 5}
}
\value{
updated icd-10-diag tables
}
\description{
run a full refresh of the icd-10 scraper for diagnosiscodes
note, this takes a long time due to the sleep time in between http requests
this is to guarantee we do not ddos the website of interest
}
\examples{
# diag_scrape(5)
}
