% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icdR.R
\name{find_icd}
\alias{find_icd}
\title{Find any relevant icd-8 or icd-10 codes}
\usage{
find_icd(
  icd_data = icd_data,
  icd_code = icd_code,
  regex = NULL,
  negate = FALSE,
  is_icd_8_code = FALSE,
  regex_icd_8 = NULL,
  regex_icd_10 = NULL
)
}
\arguments{
\item{icd_code}{is the column (variable) with icd codes}

\item{regex}{is the regex of icd codes}

\item{is_icd_8_code}{is the indicator of whether the code is icd-8 code}

\item{icd}{is the data with icd codes}
}
\value{
the dataframe with icd-8 codes
}
\description{
Finds specific icd codes provided as regex
}
\examples{
find_icd(regex = "^DE10|^DE11|^DH360|^249|^250")
}
