% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icdR.R
\name{find_icd10}
\alias{find_icd10}
\title{Find relevant icd-10 codes}
\usage{
find_icd10(icd_10 = icd_10, icd_10_code = icd_10_code, regex, negate = FALSE)
}
\arguments{
\item{icd_10}{is the data with icd-10 codes}

\item{icd_10_code}{is the column (variable) with icd-10 codes}

\item{regex}{is the regex of icd-10 codes}
}
\value{
the dataframe with icd-8 codes
}
\description{
Finds specific icd-10 codes provided as regex
}
\examples{
find_icd10(regex = "^DE10|^DE11|^DH360")
}
