% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/billtype.R
\name{get_caretype}
\alias{get_caretype}
\title{Extract Care-Type information from Billtype}
\usage{
get_caretype(btcode)
}
\arguments{
\item{btcode}{a billtype code or vector of codes}
}
\value{
a string or vector of strings with care-type info
}
\description{
Extract Care-Type information from Billtype
}
\examples{
get_caretype(c('0123','731'))
}
