% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/billtype.R
\name{get_facility}
\alias{get_facility}
\title{Extract Facility information from Billtype}
\usage{
get_facility(btcode)
}
\arguments{
\item{btcode}{a billtype code or vector of codes}
}
\value{
a string or vector of strings with facility info
}
\description{
Extract Facility information from Billtype
}
\examples{
get_facility(c('0123','731'))
}
