% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/billtype.R
\name{get_frequency}
\alias{get_frequency}
\title{Extract Frequency information from Billtype}
\usage{
get_frequency(btcode)
}
\arguments{
\item{btcode}{a billtype code or vector of codes}
}
\value{
a string or vector of strings with frequency info
}
\description{
Extract Frequency information from Billtype
}
\examples{
get_frequency(c('0123','731'))
}
