% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/placeofservice.R
\name{get_pos_desc}
\alias{get_pos_desc}
\title{Extract POS description}
\usage{
get_pos_desc(poscode)
}
\arguments{
\item{poscode}{a placeofservice code or vector of codes}
}
\value{
a vector of corresponding pos descriptions
}
\description{
Extract POS description
}
\examples{
get_pos_desc(c('11','71'))
}
