% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/placeofservice.R
\name{get_pos_name}
\alias{get_pos_name}
\title{Extract POS name}
\usage{
get_pos_name(poscode)
}
\arguments{
\item{poscode}{a placeofservice code or vector of codes}
}
\value{
a vector of corresponding pos names
}
\description{
Extract POS name
}
\examples{
get_pos_name(c('11','71'))
}
