% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/revenue.R
\name{get_rev_desc}
\alias{get_rev_desc}
\title{Extract Revenuecode code description}
\usage{
get_rev_desc(rvcode)
}
\arguments{
\item{rvcode}{a revenuecode code or vector of codes}
}
\value{
a string or vector of strings with description info
}
\description{
Extract Revenuecode code description
}
\examples{
get_rev_desc(c('0450','0111'))
}
