% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/revenue.R
\name{get_rev_group}
\alias{get_rev_group}
\title{Extract Revenuecode group description}
\usage{
get_rev_group(rvcode)
}
\arguments{
\item{rvcode}{a revenuecode code or vector of codes}
}
\value{
a string or vector of strings with group description info
}
\description{
Extract Revenuecode group description
}
\examples{
get_rev_group(c('0450','0111'))
}
