% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/revenue.R
\name{get_rev_parent}
\alias{get_rev_parent}
\title{Extract Revenuecode parent code category}
\usage{
get_rev_parent(rvcode)
}
\arguments{
\item{rvcode}{a revenuecode code or vector of codes}
}
\value{
a string or vector of strings with parent info
}
\description{
For example, the parent of 0451 is 045X, where x is a filler character
}
\examples{
get_rev_parent(c('0450','0111'))
}
