% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{icd10map}
\alias{icd10map}
\title{default map}
\format{
`icd10map`
a dataset with an example of the icd_scrape functionality
\describe{
 \item{idk}{test}
 \item {more testing, is this working}{shrug, probably not}
 \item {pe}{nis}
 ...
}
}
\source{
bozo
}
\usage{
icd10map
}
\description{
a icd-map generated in 2022
}
\keyword{datasets}
