% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icd_scrape.R
\name{icd_clean}
\alias{icd_clean}
\title{remove the icd from a string to leave just descriptors}
\usage{
icd_clean(vec, proc = F)
}
\arguments{
\item{vec}{a vector of strings}

\item{proc}{indicator for procedurecode scrape}
}
\value{
trimmed vector
}
\description{
remove the icd from a string to leave just descriptors
}
\examples{
icd_clean(c("U07.1 Covid", "E11 Diabetes"))
}
