% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnosis.R
\name{icd_trim}
\alias{icd_trim}
\title{Trim ICD-10 codes}
\usage{
icd_trim(vec, proc = F)
}
\arguments{
\item{vec}{a vector of strings}

\item{proc}{indicator for procedurecode scrape, currently unused}
}
\value{
trimmed vector
}
\description{
remove ICD-10 codes from a string
}
\examples{
icd_trim(c("U07.1 Covid", "E11 Diabetes"))
}
