% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icdr_sample.R
\name{icdr_sample}
\alias{icdr_sample}
\title{Generate some sample codes}
\usage{
icdr_sample(n, code = "diag")
}
\arguments{
\item{n}{number of codes wanted}

\item{code}{type of code desired}
}
\value{
a random sample of the respective code type
}
\description{
Generate some sample codes
}
\examples{
icdr_sample(10)

}
