% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{placeofservice}
\alias{placeofservice}
\title{placeofservice maps}
\format{
`placeofservice table`
a dataset to facilitate the categorization of placeofservice codes
}
\source{
https://www.cms.gov/Medicare/Coding/place-of-service-codes/Place_of_Service_Code_Set
}
\usage{
placeofservice
}
\description{
a map for placeofservice details
}
\details{
as of May 7th 2023
}
\keyword{datasets}
