% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/placeofservice.R
\name{pos_check}
\alias{pos_check}
\title{Validate placeofservice}
\usage{
pos_check(poscode)
}
\arguments{
\item{poscode}{a placeofservice code or vector of codes}
}
\value{
boolean for validity
}
\description{
check if placeofservice code is valid
}
\examples{
pos_check('11')
}
