% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icdR.R
\name{read_atc}
\alias{read_atc}
\title{Reads the Anatomical Therapeutic Chemical (ATC) Classification codes}
\usage{
read_atc(
  atc_link = "https://medstat.dk/da/download/file/YXRjX2NvZGVfdGV4dC50eHQ="
)
}
\arguments{
\item{atc_link}{is the link to medstat.dk}
}
\value{
the dataframe with ATC codes
}
\description{
Reads the Anatomical Therapeutic Chemical (ATC) Classification codes from medstat.dk (not from https://www.whocc.no/atc_ddd_index/).The available data include the ATC code column, the name of the active drug agent, the unit (DDD or other; not availbale for every drug)
}
\examples{
atc <- read_atc()
}
