% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icdR.R
\name{read_icd10}
\alias{read_icd10}
\title{Reads ICD-10 codes and NOMESCO procedure codes}
\usage{
read_icd10(
  sks_link = "ftp://filer.sst.dk/filer/sks/data/skscomplete/SKScomplete.txt"
)
}
\arguments{
\item{sks_link}{is the link to SKS-browser}
}
\value{
the dataframe with ICD-10 and NOMESCO procedure codes used in Danish healthcare system
}
\description{
Reads ICD-10 codes and NOMESCO procedure codes from 1995 from SKS-browser https://medinfo.dk/sks/brows.php. The columns include the ICD-10 code, diagnosis in Danish, the date the code was introduced for use, the date the code was changed, the date the code usage was ceased, and SKS version
}
\examples{
icd10 <- read_icd10()
}
