% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icdR.R
\name{read_icd8}
\alias{read_icd8}
\title{Reads ICD-8 codes}
\usage{
read_icd8(
  sks_link = "ftp://filer.sst.dk/filer/sks/data/skscomplete/SGDklass_ICD8.txt"
)
}
\arguments{
\item{sks_link}{is the link to SKS-browser}
}
\value{
the dataframe with ICD-10 and NOMESCO procedure codes used in Danish healthcare system
}
\description{
Reads ICD-8 codes from SKS-browser https://medinfo.dk/sks/brows.php. The columns include the ICD-8 code, diagnosis in Danish or Latin, the date the code was introduced for use, the date the code was changed, the date the code usage was ceased
}
\examples{
icd8 <- read_icd8()
}
