% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icdR.R
\name{read_opr}
\alias{read_opr}
\title{Reads procedure/operations codes}
\usage{
read_opr(
  sks_link = "ftp://filer.sst.dk/filer/sks/data/skscomplete/OPRklass_1995.txt"
)
}
\arguments{
\item{sks_link}{is the link to SKS-browser}
}
\value{
the dataframe with procedure codes before year 1995 used in Danish healthcare system
}
\description{
Reads procedure/operations codes from SKS-browser before 1995 https://medinfo.dk/sks/brows.php. The columns include the procedure code, name of the procedure in Danish, the date the code was introduced for use, the date the code was changed, the date the code usage was ceased
}
\examples{
opr <- read_opr()
}
