% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/revenue.R
\name{revcode_check}
\alias{revcode_check}
\title{Validate revenuecode}
\usage{
revcode_check(rvcode)
}
\arguments{
\item{rvcode}{a revenuecode code or vector of codes}
}
\value{
boolean for validity
}
\description{
check if revenuecode code is valid
}
\examples{
revcode_check('0450')
}
