% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{revenuecode}
\alias{revenuecode}
\title{revenuecode maps}
\format{
`revenuecode table`
a dataset to facilitate the categorization of revenuecode
}
\source{
https://med.noridianmedicare.com/web/jea/topics/claim-submission/revenue-codes
}
\usage{
revenuecode
}
\description{
a map for revenuecode details
}
\details{
as of May 1st 2023
}
\keyword{datasets}
