% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{icd10map}
\alias{icd10map}
\title{default map}
\format{
`icd10map table`
a dataset with an example of the icd_scrape functionality
}
\source{
bozo
}
\usage{
icd10map
}
\description{
a table for categorizing diagnosiscodes
}
\details{
as of May 3rd 2023
}
\keyword{datasets}
