% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnosis.R
\name{icd_check}
\alias{icd_check}
\title{Validate ICD-10}
\usage{
icd_check(code, dataset = NULL, column = "l4c")
}
\arguments{
\item{code}{a vector of strings}

\item{dataset}{name of dataframe, default uses the internal table}

\item{column}{name of column, default uses the relevant columns in the internal table}
}
\value{
boolean
}
\description{
check if ICD-10 code is in the database
}
\examples{
icd_check('E11')
}
