% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icd_scrape.R
\name{icd_extract}
\alias{icd_extract}
\title{extract the icd from a string to leave just descriptors}
\usage{
icd_extract(vec)
}
\arguments{
\item{vec}{a vector of strings}
}
\value{
trimmed vector
}
\description{
extract the icd from a string to leave just descriptors
}
\examples{
icd_clean(c("U07.1 Covid", "E11 Diabetes"))
}
